% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups.get}
\alias{groups.get}
\title{Gets one resource by id.}
\usage{
groups.get(
  groupUniqueId,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["groupsSettings.groups.get"]]$scopes),
  return_response = F
)
}
\arguments{
\item{groupUniqueId}{The group's email address.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/apps.groups.settings
}
}
\seealso{
\href{https://developers.google.com/google-apps/groups-settings/get_started}{Google Documentation}
}
